refactor(api): split up protocol engine types
This was getting _quite_ unwieldy. Let's split it into multiple files.
We can change what the names of the files are later if we want because
nothing is poking through and importing them individually (except each other).
